chore: restructure folders and namespaces for consistency and maintainability
- Moved services into dedicated folders (e.g., `File`, `Dialog`, `Toasts`, `Launch`) and updated namespaces accordingly.
- Organized view models by adding regions for fields, properties, services, and commands.
- Refined XAML templates by replacing `SecondaryColorTemplate` with `TapBarTemplate`.
- Simplified styles by removing redundant comments and consolidated resources in `DarkTheme.xaml`.
- Adjusted test setups to reflect updated namespaces and ensure correctness.
