+++ Kommunist.Application/Controls/MyTabBar.xaml.cs
@@ -1,6 +1,7 @@
 using System.Windows.Input;
 using Kommunist.Application.Services;
 using Kommunist.Application.Services.Navigation;
+using Kommunist.Application.Services.Toasts;
 using Kommunist.Core.Services.Interfaces;
 
 namespace Kommunist.Application.Controls;
@@ -19,7 +20,6 @@ public partial class MyTabBar
         init => SetValue(NavigateToCommandProperty, value);
     }
 
-    // Default constructor used by XAML/runtime
     public MyTabBar()
         : this(new MauiNavigationService(new MauiShellNavigator()), new MauiToastService(), initializeComponent: true)
     {
