+++ Kommunist.Application/MauiProgram.cs
@@ -3,6 +3,10 @@ using System.Reflection;
 using CommunityToolkit.Maui;
 using Kommunist.Application.Helpers;
 using Kommunist.Application.Services;
+using Kommunist.Application.Services.Dialog;
+using Kommunist.Application.Services.File;
+using Kommunist.Application.Services.Launch;
+using Kommunist.Application.Services.Toasts;
 using Kommunist.Application.ViewModels;
 using Kommunist.Core.Config;
 using Kommunist.Core.Services;
@@ -68,7 +72,6 @@ public static class MauiProgram
         builder
             .UseMauiApp<App>()
             .UseMauiCommunityToolkit()
-            // .UseMauiCommunityToolkitMarkup()
             .ConfigureFonts(ConfigureFonts);
     }
