+++ Kommunist.Application/Resources/Styles/ApplicationStyles.xaml
@@ -35,7 +35,7 @@
         <Setter Property="FontAttributes" Value="Bold" />
     </Style>
     
-    <ControlTemplate x:Key="SecondaryColorTemplate">
+    <ControlTemplate x:Key="TapBarTemplate">
         <Grid RowDefinitions="*,Auto">
             <ContentPresenter Grid.Row="0" />
             <Grid Grid.Row="1" VerticalOptions="End">
@@ -108,31 +108,11 @@
         <Setter Property="BackgroundColor" Value="{DynamicResource PageBackgroundColor}" />
     </Style>
 
-    <!-- <Style TargetType="{x:Type Label}"> -->
-    <!--     <Setter Property="TextColor" Value="{DynamicResource ContentTextColor}" /> -->
-    <!--     <Setter Property="FontSize" Value="{StaticResource SmallFontSize}" /> -->
-    <!-- </Style> -->
-
-    <!-- <Style TargetType="{x:Type Entry}"> -->
-    <!--     <Setter Property="TextColor" Value="{DynamicResource ContentTextColor}" /> -->
-    <!--     <Setter Property="FontSize" Value="{StaticResource SmallFontSize}" /> -->
-    <!-- </Style> -->
-
     <Style TargetType="{x:Type Editor}">
         <Setter Property="TextColor" Value="{DynamicResource ContentTextColor}" />
         <Setter Property="FontSize" Value="{StaticResource SmallFontSize}" />
         <Setter Property="AutoSize" Value="TextChanges" />
     </Style>
-
-    <!-- <Style TargetType="{x:Type Picker}"> -->
-    <!--     <Setter Property="TextColor" Value="{DynamicResource ContentTextColor}" /> -->
-    <!--     <Setter Property="FontSize" Value="{StaticResource SmallFontSize}" /> -->
-    <!-- </Style> -->
-
-    <!-- <Style TargetType="{x:Type Switch}"> -->
-    <!--     <Setter Property="OnColor" Value="{DynamicResource PrimaryColor}" /> -->
-    <!--     <Setter Property="ThumbColor" Value="#EEEEEF" /> -->
-    <!-- </Style> -->
 
     <Style TargetType="{x:Type Grid}">
         <Setter Property="ColumnSpacing" Value="0" />
@@ -353,10 +333,10 @@
     </Style>
     
     <Style x:Key="LabelStyle" TargetType="Label">
-        <!-- <Setter Property="TextColor" Value="{DynamicResource TextColor}"/> -->
-        <!-- <Setter Property="FontSize" Value="16"/> -->
-        <!-- <Setter Property="FontAttributes" Value="Bold"/> -->
-        <!-- <Setter Property="Margin" Value="5,0,0,5"/> -->
+        <Setter Property="TextColor" Value="{DynamicResource TextColor}"/>
+        <Setter Property="FontSize" Value="16"/>
+        <Setter Property="FontAttributes" Value="Bold"/>
+        <Setter Property="Margin" Value="5,0,0,5"/>
     </Style>
             
     <Style x:Key="CheckboxLabelStyle" TargetType="Label">
