+++ Kommunist.Application/Services/File/FileSystemService.cs
@@ -1,7 +1,7 @@
 using System.Text;
 using Kommunist.Core.Services.Interfaces;
 
-namespace Kommunist.Application.Services;
+namespace Kommunist.Application.Services.File;
 
 public class FileSystemService(string appDataDirectory) : IFileSystemService
 {
