+++ Kommunist.Application/Themes/DarkTheme.xaml
@@ -4,7 +4,6 @@
                     xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
                     x:Class="Kommunist.Application.Themes.DarkTheme">
 
-    <!-- Primary Colors -->
     <Color x:Key="PrimaryColor">#82B1FF</Color>
     <Color x:Key="PrimaryDarkColor">#8C9EFF</Color>
     <Color x:Key="Secondary">#2A2A2E</Color>
@@ -12,7 +11,6 @@
     <Color x:Key="Success">#66BB6A</Color>
     <Color x:Key="Primary">#512BD4</Color>
 
-    <!-- Text Colors -->
     <Color x:Key="ContentTextColor">White</Color>
     <Color x:Key="TextPrimaryColor">#EDEDED</Color>
     <Color x:Key="TextSecondaryColor">#B0B0B0</Color>
@@ -20,17 +18,14 @@
     <Color x:Key="TextColor">#DDDDDD</Color>
     <Color x:Key="TitleTextColor">#DDDDDD</Color>
 
-    <!-- Background & Surface -->
     <Color x:Key="PageBackgroundColor">#121212</Color>
     <Color x:Key="BackgroundColor">#1E1E1E</Color>
     <Color x:Key="SurfaceColor">Black</Color>
     <Color x:Key="ShadowColor">#66000000</Color>
 
-    <!-- Basic Colors -->
     <Color x:Key="White">#FFFFFF</Color>
     <Color x:Key="Black">#000000</Color>
 
-    <!-- Gray Scale (inverted) -->
     <Color x:Key="Gray100">#1E1E1E</Color>
     <Color x:Key="Gray200">#2C2C2C</Color>
     <Color x:Key="Gray300">#3A3A3A</Color>
