+++ Kommunist.Application/XCalendar/EventDay.cs
