+++ Kommunist.Tests/Services/LauncherServiceTests.cs
@@ -1,4 +1,5 @@
 using Kommunist.Application.Services;
+using Kommunist.Application.Services.Launch;
 using Moq;
 
 namespace Kommunist.Tests.Services;
