+++ Kommunist.Tests/Services/MauiNavigationServiceTests.cs
@@ -1,5 +1,6 @@
 using Kommunist.Application.Services;
 using Kommunist.Application.Services.Navigation;
+using Kommunist.Core.Services.Interfaces.Shared;
 using Moq;
 
 namespace Kommunist.Tests.Services;
